Improvements

+++ ffmpeg-concat.sh
@@ -17,10 +17,13 @@
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 
-# Defaults arguments
+# Arguments
 BANNER=$1
+DIRECTORY=${2%/}
+NAME=$3
+
+# Options defaults
 SCALE="640x360"
-NAME="complete"
 FORMAT="mp4"
 LOGLEVEL="info"
 
@@ -32,19 +35,18 @@ This program is free software: GPL v3 or any later version
 Copyright (C) 2015 Rosario Prestigiacomo
 
 Usage:
-ffmpeg-concat <banner> [options]
+ffmpeg-concat <banner> <video-directory> <output-filename> [options]
 
 Options:
 -s    scale - set the resolution (default is 640x360)
--n    name - complete video filename will be original-filename + name + format (default is complete)
 -f    format - select the output format (default mp4)
 -l    loglevel - log level for ffmpeg (default info)
 -h    help - show this dialog
 
 EOM
 
-# Check if the first argument is empty
-if [ "$1" == '' ]; then
+# Check the arguments
+if [ $# -lt 3 ]; then
     echo "$USAGE"
     exit 1
 fi
@@ -55,23 +57,26 @@ if [ ! -f "$BANNER" ]; then
     exit 1
 fi
 
+# Check video dir exists
+if [ ! -d "$DIRECTORY" ]; then
+    echo "Video directory - $DIRECTORY - not found"
+    exit 1
+fi
+
 # Getopts options
-shift
-while getopts ':hs:n:f:l:' OPTION; do
+while getopts ':hrkns:f:l:' OPTION; do
     case "$OPTION" in
         h) echo "$USAGE"
             exit 1
             ;;
         s) SCALE=$OPTARG
             ;;
-        n) NAME=$OPTARG
-            ;;
         f) FORMAT=$OPTARG
             ;;
         l) LOGLEVEL=$OPTARG
             ;;
         :) printf "Missing argument for -%s\n" "$OPTARG" >&2
-           echo "$usage" >&2
+           echo "$USAGE" >&2
            exit 1
            ;;
         \?) printf "Illegal option: -%s\n" "$OPTARG" >&2
@@ -80,7 +85,7 @@ while getopts ':hs:n:f:l:' OPTION; do
            ;;
     esac
 done
-#shift $((OPTIND-1))
+shift $((OPTIND - 3))
 
 # Colors
 g=$'\e[32m'
@@ -94,22 +99,33 @@ ERRNUM=1
 SKIPPED[0]=$y"Skipped:"$c
 SKIPNUM=1
 
+# Resize the banner to the default or desired resolution
+ffmpeg -i "$BANNER" -loglevel "$LOGLEVEL" -strict -2 -vf scale="$SCALE",setsar=1:1 -c:v libx264 -preset fast -profile:v main -crf 20 "$DIRECTORY"/"$BANNER"-banner-resize."$FORMAT"
+if [ $? -eq 0 ]; then
+    RESBAN="$DIRECTORY"/"$BANNER"-banner-resize."$FORMAT"
+else
+    # Log banner resize error
+    ERRORS[ERRNUM]="Cannot resize - $BANNER - banner."
+    ((ERRNUM++))
+fi
+
+
 # For every file in the directory
-for VIDEO in *
+for VIDEO in "$DIRECTORY"/*
 do
     # Check if the file is a video(doesn't work with mkv)
     if file -i "$VIDEO" | grep -q video ; then
         # Don't execute if the file is the banner or the script itself
-        if [ "$VIDEO" != "$BANNER" ] && [ "$VIDEO" != "$0" ]; then
+        if [ "$VIDEO" != "$BANNER" ] && [ "$VIDEO" != "$RESBAN" ] && [ "$VIDEO" != "$0" ]; then
             # Resize video
             ffmpeg -i "$VIDEO" -loglevel "$LOGLEVEL" -strict -2 -vf scale="$SCALE",setsar=1:1 -c:v libx264 -preset fast -profile:v main -crf 20 "$VIDEO"-resize."$FORMAT"
             # In case of success
             if [ $? -eq 0 ]; then
                 # Create the concat
-                ffmpeg -i "$BANNER" -i "$VIDEO"-resize."$FORMAT" -i "$BANNER" -loglevel "$LOGLEVEL" -strict -2 -filter_complex "[0:0] [0:1] [1:0] [1:1] [2:0] concat=n=3:v=1:a=1 [v] [a]" -map "[v]" -map "[a]" "$VIDEO"-"$NAME"."$FORMAT"
+                ffmpeg -i "$RESBAN" -i "$VIDEO"-resize."$FORMAT" -i "$RESBAN" -loglevel "$LOGLEVEL" -strict -2 -filter_complex "[0:0] [0:1] [1:0] [1:1] [2:0] concat=n=3:v=1:a=1 [v] [a]" -map "[v]" -map "[a]" "$VIDEO"-"$NAME"."$FORMAT"
                 # If everithing was fine and dandy, clean and move on
                 if [ $? -eq 0 ]; then
-                    echo ""
+                    printf '%s\n'
                     echo "Cleaning - $VIDEO - resize temp file."
                     rm "$VIDEO"-resize."$FORMAT"
                 else
@@ -130,16 +146,14 @@ do
     fi
 done
 
-echo ""
+printf '%s\n'
 
 if [ "$SKIPNUM" != 1 ]; then
-    printf '%s\n' "${SKIPPED[@]}"
-    echo ""
+    printf '%s\n' "${SKIPPED[@]}" '%s\n'
 fi
 
 if [ "$ERRNUM" != 1 ]; then
-    printf '%s\n' "${ERRORS[@]}"
-    echo ""
+    printf '%s\n' "${ERRORS[@]}" '%s\n'
 fi
 
 # Final report
